Invariant specific implementations for UTF-8 and UTF-16; removed old versions; code structure cleanup
